Front: Move Registration Methods to Tasks Service
